requirements.txt
PyQt6>=6.6.0
matplotlib>=3.8.0
python-docx>=1.1.0
reportlab>=4.0.0
